Rename package used in scala example.
Rename examples/src/scala/com/pants/example to examples/src/scala/compants/examples to mirror the
convention used in the Java examples.
