+++ examples/src/java/com/pants/examples/java_sources/BUILD
@@ -4,6 +4,6 @@
 java_library(name = 'java_sources',
   sources = globs('*.java'),
   dependencies = [
-   'examples/src/scala/com/pants/example/scala_with_java_sources',
+   'examples/src/scala/com/pants/examples/scala_with_java_sources',
   ],
 )
