+++ examples/src/java/com/pants/examples/java_sources/Greet.java
