+++ examples/src/java/com/pants/examples/jaxb/reader/VegetableReader.java
