+++ examples/src/resources/com/pants/examples/hello/BUILD
