+++ examples/src/resources/com/pants/examples/hello/world.txt
