+++ examples/src/resources/com/pants/examples/jaxb/BUILD
