+++ examples/src/resources/com/pants/examples/jaxb/simple.xsd
