+++ examples/src/resources/com/pants/examples/names/BUILD
