+++ examples/src/resources/com/pants/examples/names/simple.xml
