+++ examples/src/scala/com/pants/examples/hello/README
