+++ examples/src/scala/com/pants/examples/hello/docs/README_DOCS
