+++ examples/src/scala/com/pants/examples/scala_with_java_sources/Greet.scala
