+++ examples/src/thrift/com/pants/examples/BUILD
