+++ tests/python/pants_test/backend/jvm/tasks/jvm_compile/java/BUILD
