+++ tests/python/pants_test/backend/jvm/tasks/jvm_compile/scala/testdata/simple_analysis_split0
@@ -25,14 +25,14 @@ name hashing:
 0 -> false
 products:
 3 items
-/src/pants/examples/src/scala/com/pants/example/hello/welcome/Welcome.scala -> /src/pants/.pants.d/compile/jvm/scala/classes/com/pants/example/hello/welcome/Welcome$$anonfun$apply$1.class
-/src/pants/examples/src/scala/com/pants/example/hello/welcome/Welcome.scala -> /src/pants/.pants.d/compile/jvm/scala/classes/com/pants/example/hello/welcome/Welcome$.class
-/src/pants/examples/src/scala/com/pants/example/hello/welcome/Welcome.scala -> /src/pants/.pants.d/compile/jvm/scala/classes/com/pants/example/hello/welcome/Welcome.class
+/src/pants/examples/src/scala/com/pants/examples/hello/welcome/Welcome.scala -> /src/pants/.pants.d/compile/jvm/scala/classes/com/pants/examples/hello/welcome/Welcome$$anonfun$apply$1.class
+/src/pants/examples/src/scala/com/pants/examples/hello/welcome/Welcome.scala -> /src/pants/.pants.d/compile/jvm/scala/classes/com/pants/examples/hello/welcome/Welcome$.class
+/src/pants/examples/src/scala/com/pants/examples/hello/welcome/Welcome.scala -> /src/pants/.pants.d/compile/jvm/scala/classes/com/pants/examples/hello/welcome/Welcome.class
 binary dependencies:
 3 items
-/src/pants/examples/src/scala/com/pants/example/hello/welcome/Welcome.scala -> /Library/Java/JavaVirtualMachines/jdk1.8.0_40.jdk/Contents/Home/jre/lib/rt.jar
-/src/pants/examples/src/scala/com/pants/example/hello/welcome/Welcome.scala -> /src/pants/.pants.d/compile/jvm/java/classes/com/pants/examples/hello/greet/Greeting.class
-/src/pants/examples/src/scala/com/pants/example/hello/welcome/Welcome.scala -> /src/pants/.pants.d/ivy/jars/org.scala-lang/scala-library/jars/scala-library-2.10.4.jar
+/src/pants/examples/src/scala/com/pants/examples/hello/welcome/Welcome.scala -> /Library/Java/JavaVirtualMachines/jdk1.8.0_40.jdk/Contents/Home/jre/lib/rt.jar
+/src/pants/examples/src/scala/com/pants/examples/hello/welcome/Welcome.scala -> /src/pants/.pants.d/compile/jvm/java/classes/com/pants/examples/hello/greet/Greeting.class
+/src/pants/examples/src/scala/com/pants/examples/hello/welcome/Welcome.scala -> /src/pants/.pants.d/ivy/jars/org.scala-lang/scala-library/jars/scala-library-2.10.4.jar
 direct source dependencies:
 0 items
 direct external dependencies:
@@ -51,19 +51,19 @@ inheritance external dependencies:
 0 items
 class names:
 3 items
-/src/pants/examples/src/scala/com/pants/example/hello/welcome/Welcome.scala -> com.pants.example.hello.welcome.Welcome
-/src/pants/examples/src/scala/com/pants/example/hello/welcome/Welcome.scala -> com.pants.example.hello.welcome.Welcome$
-/src/pants/examples/src/scala/com/pants/example/hello/welcome/Welcome.scala -> com.pants.example.hello.welcome.Welcome$$anonfun$apply$1
+/src/pants/examples/src/scala/com/pants/examples/hello/welcome/Welcome.scala -> com.pants.examples.hello.welcome.Welcome
+/src/pants/examples/src/scala/com/pants/examples/hello/welcome/Welcome.scala -> com.pants.examples.hello.welcome.Welcome$
+/src/pants/examples/src/scala/com/pants/examples/hello/welcome/Welcome.scala -> com.pants.examples.hello.welcome.Welcome$$anonfun$apply$1
 used names:
 0 items
 product stamps:
 3 items
-/src/pants/.pants.d/compile/jvm/scala/classes/com/pants/example/hello/welcome/Welcome$$anonfun$apply$1.class -> lastModified(1426908125000)
-/src/pants/.pants.d/compile/jvm/scala/classes/com/pants/example/hello/welcome/Welcome$.class -> lastModified(1426908125000)
-/src/pants/.pants.d/compile/jvm/scala/classes/com/pants/example/hello/welcome/Welcome.class -> lastModified(1426908125000)
+/src/pants/.pants.d/compile/jvm/scala/classes/com/pants/examples/hello/welcome/Welcome$$anonfun$apply$1.class -> lastModified(1426908125000)
+/src/pants/.pants.d/compile/jvm/scala/classes/com/pants/examples/hello/welcome/Welcome$.class -> lastModified(1426908125000)
+/src/pants/.pants.d/compile/jvm/scala/classes/com/pants/examples/hello/welcome/Welcome.class -> lastModified(1426908125000)
 source stamps:
 1 items
-/src/pants/examples/src/scala/com/pants/example/hello/welcome/Welcome.scala -> hash(2245c9e68457b389bac32ccaa0dca5ddadd8f1da)
+/src/pants/examples/src/scala/com/pants/examples/hello/welcome/Welcome.scala -> hash(2245c9e68457b389bac32ccaa0dca5ddadd8f1da)
 binary stamps:
 3 items
 /Library/Java/JavaVirtualMachines/jdk1.8.0_40.jdk/Contents/Home/jre/lib/rt.jar -> lastModified(1422417173000)
@@ -76,13 +76,13 @@ class names:
 /src/pants/.pants.d/ivy/jars/org.scala-lang/scala-library/jars/scala-library-2.10.4.jar -> scala.collection.TraversableLike
 internal apis:
 1 items
-/src/pants/examples/src/scala/com/pants/example/hello/welcome/Welcome.scala -> 
+/src/pants/examples/src/scala/com/pants/examples/hello/welcome/Welcome.scala -> 
 FakeApiForWelcome
 external apis:
 0 items
 source infos:
 1 items
-/src/pants/examples/src/scala/com/pants/example/hello/welcome/Welcome.scala -> 
+/src/pants/examples/src/scala/com/pants/examples/hello/welcome/Welcome.scala -> 
 FakeSourceInfoForWelcome
 compilations:
 0 items
