+++ tests/python/pants_test/tasks/test_depmap_integration.py
@@ -66,7 +66,7 @@ class DepmapIntegrationTest(PantsRunIntegrationTest):
 
   def test_dep_map_for_java_sources(self):
     with temporary_dir(root_dir=self.workdir_root()) as workdir:
-      test_target = 'examples/src/scala/com/pants/example/scala_with_java_sources'
+      test_target = 'examples/src/scala/com/pants/examples/scala_with_java_sources'
       json_data = self.run_depmap_project_info(test_target, workdir)
       targets = json_data.get('targets')
       self.assertIn('examples/src/java/com/pants/examples/java_sources:java_sources', targets)
